+++ src/init.lua
@@ -1,6 +1,7 @@
 local DoublyLinkedList = require(script.DoublyLinkedList)
 local LinkedList = require(script.LinkedList)
 local Map = require(script.Map)
+local OrderedMap = require(script.OrderedMap)
 local OrderedSet = require(script.OrderedSet)
 local Queue = require(script.Queue)
 local Set = require(script.Set)
@@ -10,6 +11,7 @@ return {
 	DoublyLinkedList = DoublyLinkedList,
 	LinkedList = LinkedList,
 	Map = Map,
+	OrderedMap = OrderedMap,
 	OrderedSet = OrderedSet,
 	Queue = Queue,
 	Set = Set,